+++ array/zfp/iterator2.h
@@ -15,7 +15,7 @@ public:
   const_iterator() : const_handle(0, 0, 0) {}
 
   // constructor
-  explicit const_iterator(container_type* container, uint i, uint j) : const_handle(container, i, j) {}
+  explicit const_iterator(container_type* container, size_t i, size_t j) : const_handle(container, i, j) {}
 
   // dereference iterator
   const_reference operator*() const { return const_reference(container, i(), j()); }
@@ -29,8 +29,8 @@ public:
   const_iterator operator++(int) { const_iterator it = *this; increment(); return it; }
 
   // container indices of value referenced by iterator
-  uint i() const { return const_handle::i; }
-  uint j() const { return const_handle::j; }
+  size_t i() const { return const_handle::i; }
+  size_t j() const { return const_handle::j; }
 
 protected:
   void increment()
@@ -70,7 +70,7 @@ public:
   iterator() : const_iterator(0, 0, 0) {}
 
   // constructor
-  explicit iterator(container_type* container, uint i, uint j) : const_iterator(container, i, j) {}
+  explicit iterator(container_type* container, size_t i, size_t j) : const_iterator(container, i, j) {}
 
   // dereference iterator
   reference operator*() const { return reference(container, i(), j()); }
